backend/api/nft/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
jinja2